Major refactoring, renaming modules
